src_fast_api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy>=2.0.36
psycopg[binary]>=3.2.2,<4.0.0
python-dotenv==1.0.0
email-validator>=2.0.0

